Changes to ArenaPatch, makes it so you can have additional crowd or signs (Looks for PNG files in crowd or signs respectively and adds them to list of ones to randomly use)
Also changes way it applies signs and crowd textures in custom arenas so it keeps the custom arena shaders etc to avoid weird partially see through issue that happens. Wondering if a similar logic could be used to change the default material to use something from custom arena materials to avoid issue on wrestlers themselves when on a custom arena?